Style changes
Div over the video for play/pause action is now 100% width and height.

Play/Pause animation is shown only when the video is loaded/running.

Fixed bug of the weird play/pause overlapping button.

Buffering text re-positioned. Now it doesn't displays a vertical scrollbar.

Semifading appname and title to 50% so it's not that invasive for the user